+++ src/index.tsx
@@ -1,12 +1,12 @@
 import { SnackbarProvider } from "notistack";
 import { render } from "react-dom";
-import Main from "./Main";
+import App from "./App";
 
 const rootElement = document.getElementById("root");
 
 render(
   <SnackbarProvider maxSnack={1} autoHideDuration={2000}>
-    <Main />
+    <App />
   </SnackbarProvider>,
   rootElement
 );
